Fix countdown off-by-one, iOS background audio, splash replay on rotation
- RunnerEngine now emits countdown cues using ceiling seconds so voice
  announcements line up with the ProgressRing digit on screen.
- IosRunnerForegroundController plays a silent AVAudioEngine loop while the
  timer is active so iOS doesn't suspend the process between cues; voice
  countdown now continues when the app is backgrounded.
- Splash converted from a NavHost route into an overlay composable. On iOS
  the saveable-state-registry resets on trait changes (rotation), which
  recreated the NavController and re-played the typewriter; the overlay is
  driven by a singleton AppViewModel flag that survives that reset. AppEvent
  Navigation was unused and removed.

+++ features/timers/impl/src/iosMain/kotlin/com/hiittimer/features/timers/impl/audio/IosAudioCuePlayerFactory.kt
@@ -102,7 +102,7 @@ class IosAudioCuePlayer(
         private const val LONG_BEEP: UInt = 1052u
         private const val FINISH_SOUND: UInt = 1025u
 
-        private const val MIX_WITH_OTHERS: UInt = 1u
-        private const val DUCK_OTHERS: UInt = 2u
+        private const val MIX_WITH_OTHERS: ULong = 1u
+        private const val DUCK_OTHERS: ULong = 2u
     }
 }
